Add support for redirection
Can now successfully redirect commands to and from file, note that file
redirection takes precedence over piping, so if you execute a commmand
like this: 'cat names.text > new.txt | sort', the names in names.txt
will not reach 'sort' and will not be sorted, the redirection to file
'new.txt' takes precedence. This is similar behaviour to bash.