Reap zombie processes Stop SIGCHLD garbling input
Added a signal handler to ensure that zombie processes spawned by the
shell are reaped.

Prevented the SIGCHLD signal handler from garbling the user's input when
a background process returns.

Also added a critical section to the code that executes and waits for a
sequential command to return. This prevents the 'waitpid' function call
from failing if the SIGCHLD handler is called while it is waiting.